+++ Scripts/Engines/BulkOrders/LargeBOD.cs
@@ -218,13 +218,13 @@ namespace Server.Engines.BulkOrders
 		{
 			base.Serialize( writer );
 
-			writer.Write( (int) 0 ); // version
+			writer.Write( 0 ); // version
 
 			writer.Write( m_AmountMax );
 			writer.Write( m_RequireExceptional );
 			writer.Write( (int) m_Material );
 
-			writer.Write( (int) m_Entries.Length );
+			writer.Write( m_Entries.Length );
 
 			for ( int i = 0; i < m_Entries.Length; ++i )
 				m_Entries[i].Serialize( writer );
