+++ Scripts/Items/Misc/PromotionalToken.cs
@@ -58,7 +58,7 @@ namespace Server.Items
 		{
 			base.Serialize( writer );
 
-			writer.Write( (int)0 );
+			writer.Write( 0 );
 		}
 
 		public override void Deserialize( GenericReader reader )
@@ -143,7 +143,7 @@ namespace Server.Items
 		{
 			base.Serialize( writer );
 
-			writer.Write( (int)0 );
+			writer.Write( 0 );
 		}
 
 		public override void Deserialize( GenericReader reader )
